renamed the action and reducer files
