+++ code-kata/src/reducers/catfish-reducer.js
